+++ hypervisor/arch/x86/exception.c
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <asm/cpu.h>
 #include <asm/irq.h>
 #include <debug/dump.h>
+#include <cpu.h>
 
 void dispatch_exception(struct intr_excp_ctx *ctx)
 {
