+++ hypervisor/arch/x86/hw_thermal.c
@@ -8,7 +8,7 @@
 #include <softirq.h>
 #include <irq.h>
 #include <logmsg.h>
-#include <asm/cpu.h>
+#include <cpu.h>
 #include <asm/msr.h>
 #include <asm/irq.h>
 #include <asm/apicreg.h>
